refactor: extract retry policy constants into configurable RetryConfig
